Move DropGrammar to `grammar` package
