+++ src/main/java/org/terasology/drops/grammar/ParseResult.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.terasology.drops;
+package org.terasology.drops.grammar;
 
 /**
  * Denotes a drop parsed from a drop definition string.
